+++ CHANGELOG.md
@@ -34,3 +34,5 @@ Changes:
 * Add `trait AutoBumpyEntry`, to simplify entries that know their own
   size/index [#12]
 * Change error handling to use `SimpleResult` [#13]
+* Change the index..size stuff to use `std::ops::Range` [#14]
+  * Note that this breaks compatibility